backend/requirements.txt
fastapi
uvicorn
websockets
httpx
python-dotenv
pydantic
